a2abase_cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a2abase-cli"
version = "0.1.0"
description = "Production-ready CLI tool for scaffolding A2ABase Agent SDK projects"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "BelArabyAI"}
]
license = {text = "MIT"}
keywords = ["a2abase", "cli", "scaffold", "agent", "sdk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "jinja2>=3.1.4",
    "watchfiles>=0.24.0",
    "pydantic>=2.8.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.6.0",
]
ngrok = [
    "pyngrok>=7.0.0",
]

[project.scripts]
a2abase = "a2abase_cli.cli:app"

[project.urls]
Homepage = "https://github.com/A2ABaseAI/sdks"
Repository = "https://github.com/A2ABaseAI/sdks"
Issues = "https://github.com/A2ABaseAI/sdks/issues"

[tool.hatchling.build]
packages = ["src/a2abase_cli"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

